Added more finite detail to the level
Playground set has some semi-basic colors to it that match what a playground should look like (items without textures are due to them being replaced by better models to fit with the design of the level). Also added more texture to the ground and areas where items will be placed in the future. Still need to make parking lot texture and stop sign for the road and potentially others based on what is still left.